# docs/openapi.yaml
openapi: 3.1.0
info:
  title: VIVA Management Portal API
  version: 1.0.0
  description: |
    REST API for the VIVA Digital Center management portal.

    Authenticate with `POST /token`, then send the returned token as
    `Authorization: Bearer <token>`. Access to protected operations also depends
    on the staff member's role and permission overrides.
  contact:
    name: VIVA Digital Center
    email: vivadigitalcenter@gmail.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local backend
  - url: http://{host}:8000/api/v1
    description: Local network backend
    variables:
      host:
        default: localhost
        description: Hostname or LAN IP address running Laravel
...
security:
  - bearerAuth: []
paths:
  /token:
    post:
      tags: [Authentication]
      summary: Log in and issue a Sanctum access token
      operationId: issueToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: admin@example.com
              password: StrongPass123!
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /profile/password:
    put:
      tags: [Authentication]
      summary: Change the authenticated staff password
      description: New passwords require at least 8 characters, mixed case, a number, and a symbol.
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /finance/invoices:
    post:
      tags: [Finance]
      summary: Create a multi-enrollment draft invoice
      operationId: createInvoice
      x-permission: finance.manage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceRequest'
      responses:
        '201':
          $ref: '#/components/responses/DataResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /finance/invoices/{invoice}/issue:
    post:
      tags: [Finance]
      summary: Issue a draft invoice
      operationId: issueInvoice
      x-permission: finance.manage
      parameters:
        - $ref: '#/components/parameters/InvoiceId'
      responses:
        '200':
          $ref: '#/components/responses/DataResponse'
        '422':
          $ref: '#/components/responses/DomainError'
...
  /finance/payments:
    post:
      tags: [Finance]
      summary: Record and allocate a payment
      operationId: recordPayment
      x-permission: finance.manage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '201':
          $ref: '#/components/responses/DataResponse'
        '422':
          $ref: '#/components/responses/DomainError'
...
  /finance/payments/{payment}/cancel:
    post:
      tags: [Finance]
      summary: Cancel a payment and restore invoice balances
      operationId: cancelPayment
      x-permission: payments.cancel
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/DataResponse'
  /finance/invoices/{invoice}/cancel:
    post:
      tags: [Finance]
      summary: Cancel an invoice
      operationId: cancelInvoice
      x-permission: finance.manage
      parameters:
        - $ref: '#/components/parameters/InvoiceId'
      responses:
        '200':
          $ref: '#/components/responses/DataResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  parameters:
...
    InvoiceId:
      name: invoice
      in: path
      required: true
      schema: { type: integer }
...
    PaymentId:
      name: payment
      in: path
      required: true
      schema: { type: integer }
...
  responses:
    Success:
      description: Successful operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
    DataResponse:
      description: Successful operation with response data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DataResponse'
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    DomainError:
      description: Business rule prevented the operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        device_name:
          type: string
          default: web
    LoginResponse:
      type: object
      required: [success, token, user]
      properties:
        success:
          type: boolean
          example: true
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: [string, 'null'] }
        role:
          type: string
          enum: [manager, admin, facilitator]
        status:
          type: string
          enum: [pending, approved, inactive]
    ChangePasswordRequest:
      type: object
      required: [current_password, password, password_confirmation]
      properties:
        current_password:
          type: string
          format: password
        password:
          type: string
          format: password
          minLength: 8
        password_confirmation:
          type: string
          format: password
...
    InvoiceRequest:
      type: object
      required: [trainee_id, items]
      properties:
        trainee_id: { type: integer }
        due_date: { type: [string, 'null'], format: date }
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [enrollment_id, description, amount]
            properties:
              enrollment_id: { type: integer }
              description: { type: string }
              amount: { type: number, minimum: 0.01 }
    PaymentRequest:
      type: object
      required: [trainee_id, amount, method, allocations]
      properties:
        trainee_id: { type: integer }
        amount: { type: number, minimum: 0.01 }
        method:
          type: string
          enum: [cash, mobile_money]
        provider:
          type: [string, 'null']
          description: Required for mobile money.
        reference_number:
          type: [string, 'null']
          description: Required for mobile money.
        allocations:
          type: array
          minItems: 1
          items:
            type: object
            required: [invoice_id, amount]
            properties:
              invoice_id: { type: integer }
              amount: { type: number, minimum: 0.01 }
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    DataResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Endpoint-specific response data.
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string